backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
groq==0.9.0
google-generativeai==0.7.2
chromadb==0.5.5
sentence-transformers==2.7.0
transformers==4.41.2
accelerate==0.30.1
torch==2.3.0
pandas==2.2.2
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0
